Updated the pom file to include the correct activemq jar files.  I was using the 5.5.0 jar files instead of the 5.5.1 jar files.
Added the source plugin to the pom file so that the sources are installed in
the maven repo in case someone needs them to debug their code.

Removed the original hub/spoke implementation since this is no longer required.

Move the log4j.properties file from the main/resources directory to the test/resources
directory.

Removed all commented out code that was part of the original implementation
that has been shown to not be required.